#1186 - Switch for turn off/on requests for type

+++ Realization/backend/core/core-api/src/main/java/eu/bcvsolutions/idm/core/api/rest/AbstractReadWriteDtoController.java
@@ -93,6 +93,11 @@ public abstract class AbstractReadWriteDtoController<DTO extends BaseDto, F exte
 	public DTO saveDto(DTO dto, BasePermission... permission) {
 		Assert.notNull(dto, "DTO is required");
 		//
+		if (this.supportsRequests()) {
+			throw new ResultCodeException(CoreResultCode.REQUEST_CUD_OPERATIONS_NOT_ALLOWED,
+					ImmutableMap.of("controller", this.getClass().getSimpleName()));
+		}
+		//
 		return getService().save(validateDto(dto), permission);
 	}
 
@@ -221,6 +226,11 @@ public abstract class AbstractReadWriteDtoController<DTO extends BaseDto, F exte
 	public void deleteDto(DTO dto) {
 		Assert.notNull(dto, "DTO is required");
 		//
+		if (this.supportsRequests()) {
+			throw new ResultCodeException(CoreResultCode.REQUEST_CUD_OPERATIONS_NOT_ALLOWED,
+					ImmutableMap.of("controller", this.getClass().getSimpleName()));
+		}
+		//
 		getService().delete(dto, IdmBasePermission.DELETE);
 	}
 	
@@ -290,5 +300,28 @@ public abstract class AbstractReadWriteDtoController<DTO extends BaseDto, F exte
 		}
 		bulkAction.setEntityClass(getService().getEntityClass().getName());
 		bulkAction.setFilterClass(this.getFilterClass().getName());
+	}
+	
+	@Override
+	protected DTO checkAccess(DTO dto, BasePermission... permission) {
+		// If controller supports request, then only READ operation is allowed
+		if (this.supportsRequests()) {
+			if(permission != null && permission.length == 1 && IdmBasePermission.READ.equals(permission[0])) {
+				return super.checkAccess(dto, permission);
+			}
+			throw new ResultCodeException(CoreResultCode.REQUEST_CUD_OPERATIONS_NOT_ALLOWED,
+					ImmutableMap.of("controller", this.getClass().getSimpleName()));
+		}
+		return super.checkAccess(dto, permission);
+	}
+	
+	/**
+	 * If return true, then controller supports requests and cannot be used for CUD
+	 * operations. For CUD operations should be using the request controller.
+	 * 
+	 * @return
+	 */
+	protected boolean supportsRequests() {
+		return false;
 	}
 }
